refactor: compile GitLab runtime from TypeScript
