+++ gitlab/setup-vp.yml
@@ -3,12 +3,6 @@ spec:
     version:
       description: "Version of Vite+ to install"
       default: "latest"
-    node-version:
-      description: "Node.js version to install via `vp env use`."
-      default: "lts"
-    node-version-file:
-      description: "Path to file containing the Node.js version spec (.nvmrc, .node-version, .tool-versions, package.json). Takes precedence over node-version when specified."
-      default: ""
     working-directory:
       description: "Project directory used for relative paths and default `vp install` execution."
       default: "."
@@ -26,7 +20,7 @@ spec:
       description: "Optional scope for authenticating against scoped registries."
       default: ""
     setup-ref:
-      description: "setup-vp ref used to download the GitLab bootstrap and Node runtime. Pin this to the same tag or commit as the remote template for strict reproducibility."
+      description: "setup-vp ref used to download the GitLab bootstrap and compiled runtime. Pin this to the same tag or commit as the remote template for strict reproducibility."
       default: "v1"
 ---
 .setup-vp:
@@ -49,8 +43,6 @@ spec:
       }
 
       export SETUP_VP_VERSION="$[[ inputs.version ]]"
-      export SETUP_VP_NODE_VERSION="$[[ inputs.node-version ]]"
-      export SETUP_VP_NODE_VERSION_FILE="$[[ inputs.node-version-file ]]"
       export SETUP_VP_WORKING_DIRECTORY="$[[ inputs.working-directory ]]"
       SETUP_VP_RUN_INSTALL="$(cat <<'SETUP_VP_RUN_INSTALL_EOF'
       $[[ inputs.run-install ]]
